Let jQuery serialize form data
This change addresses a bug in which interesting values for secret for
instance can cause serialization failures. In my case, my application
secret contained an ampersand, `&`. Which caused the form to fail to
correctly encode the value resulting in errors.

Rather than just wrapping client_secret in `encodeURIComponent`, I opted
to migrate all the parameters to a simple object and let jQuery handle
it.